♻️ Validator 구조 개선 #9
