+++ validators/index.js
@@ -1,4 +1,4 @@
-const updateUserValidator = require('./updateUserValidator');
+const updateUserValidator = require('./users/updateUserValidator');
 
 module.exports = {
   updateUserValidator,
